Graphics... work?
So this is just a super duper bad test to see if we can read the vram from the vga controller, basically this is a direct mapping of the hack cpus memory to the screen. since the hack cpu only supports 15 bit wide addressing only the top bit gets written to and the rest is unstable. Also the reading mechanism is awful, an address gets read and we just hope it gets to the screen in time which it often doesnt so it takes 3 minutes for the screen to completely refresh (srs)